The second task of the The GRIP @TSK.
This file I have tried to explore the clustering done by K-Means. While planning and understanding the process I have also created a blog post regarding the KNN algorithm and also K- means.
The sample problem was to figure out the optimum number of clusters for the Iris dataset so I have used this case to present the dataset as an unsupervised one and explored it with K_means.